requirements.txt
telethon
pydantic>=2.12.5
python-dotenv
requests
python-dateutil

# LLM Curation
typer
rich
Pillow
langchain-groq>=0.2.4
langchain-core>=0.3.33
langchain>=0.3.15

# TTS and Video
edge-tts>=6.1.0

# Development & Testing
pytest
pytest-asyncio
pytest-mock
psutil
